updated to retrieve more metadata
Updated the spotify API retreival code to grab more artist and track metadata (including image urls) that we can possibly use for our analysis. Edited collect_random_tracks() function to store data outside the function, so if the connection is lost somehow we can still combine and export the data we retrieved up to that point.